Improved code of LightGBM to meet the specifications of tarbular
