+++ examples/models/tabular_classification/DNNTorch.py
@@ -19,6 +19,7 @@
 
 import os
 import gc
+import json
 import tempfile
 import base64
 import pickle
@@ -51,7 +52,8 @@ warnings.simplefilter(action='ignore', category=FutureWarning)
     'application_{train|test}.csv' of this competition as dataset. Some parts of
     preprocessing is based on old version `fastai` module, which is no longer
     supported and could cause imcompatible of other modules if forcibly used.
-    So a part of 'fastai.structured' is embeded in this file.
+    So a part of 'fastai.structured' is embeded in this file to ensure the model
+    runs well.
 '''
 
 ##########################
@@ -62,7 +64,7 @@ def get_sample(df, n):
     idxs = sorted(np.random.permutation(len(df))[:n])
     return df.iloc[idxs].copy()
 
-
+# for numeric data, fill up with median values
 def fix_missing(df, col, name, na_dict):
     if is_numeric_dtype(col):
         if pd.isnull(col).sum() or (name in na_dict):
@@ -81,7 +83,8 @@ def scale_vars(df, mapper):
     df[mapper.transformed_names_] = mapper.transform(df)
     return mapper
 
-
+# for category data, convert to indexes, and plus 1
+# the NA was remarked as -1 and then become 0
 def numericalize(df, col, name, max_n_cat):
     if not is_numeric_dtype(col) and ( max_n_cat is None or col.nunique()>max_n_cat):
         df[name] = col.cat.codes+1
@@ -114,8 +117,32 @@ def proc_df(df, y_fld=None, skip_flds=None, ignore_flds=None, do_scale=False, na
     return res
 
 
+##########################################
+# pytorch dataset for the home credit data
+##########################################
+
+class HomeCreditData(Dataset):
+    def __init__(self, source, target, cat_list):
+        self.x_cat = source[cat_list].values.tolist()
+        self.x_cont = source.drop(cat_list, axis=1).values.tolist()
+        self.y = None if target is None else target.values.tolist()
+        self.size = source.shape[0]
+
+    def __getitem__(self, index):
+        if self.y is not None:
+            return torch.LongTensor(self.x_cat[index]), \
+                   torch.Tensor(self.x_cont[index]), \
+                   self.y[index]
+        else:
+            return torch.LongTensor(self.x_cat[index]), \
+                   torch.Tensor(self.x_cont[index])
+
+    def __len__(self):
+        return self.size
+
+
 ####################################
-# modle receiving mixed inputs
+# This modle receiving mixed inputs
 # category and continuous data type
 ####################################
 
@@ -166,88 +193,11 @@ class MixedInputModel(nn.Module):
         x = F.log_softmax(x, dim=1)
         return x
 
-##########################################
-# pytorch dataset for the home credit data
-##########################################
-
-class HomeCreditData(Dataset):
-    def __init__(self, source, target, cat_size):
-        cat_vars = [x[0] for x in cat_size]
-        self.x_cat = source[cat_vars].values.tolist()
-        self.x_cont = source.drop(cat_vars, axis=1).values.tolist()
-        self.y = target
-        self.size = source.shape[0]
-
-    def __getitem__(self, index):
-        if self.y is not None:
-            return torch.LongTensor(self.x_cat[index]), \
-                   torch.Tensor(self.x_cont[index]), \
-                   self.y[index]
-        else:
-            return torch.LongTensor(self.x_cat[index]), \
-                   torch.Tensor(self.x_cont[index])
-
-    def __len__(self):
-        return self.size
-
-##################################
-# specific preprocessing functions
-# for train and test dataset
-##################################
-
-def load_structured_train_data(df):
-    # Optional: Remove 4 applications with XNA CODE_GENDER (train set)
-    df = df[df['CODE_GENDER'] != 'XNA']
-    # separate target
-    y = np.array(df['TARGET'])
-    df = df.drop('TARGET', axis=1)
-    # collect the names of category type columns
-    cat_names = [col for col in df.columns if df[col].dtype == 'object']
-    # calculate the size of each category. Format:[NAME, INDEX, NUM_OF_CAT]
-    cat_size = [(name, len(df[name].unique()) + 1) for name in cat_names]
-    # deal with NaN and make index encoding
-    cat_map = {}
-    for name in cat_names:
-        df[name] = df[name].astype('category').cat.as_ordered()
-        cat_codes = dict(enumerate(df[name].cat.categories))
-        cat_map[name] = {cat: code + 1 for code, cat in cat_codes.items()}
-    df_numeric, _, nas, mapper = proc_df(df, do_scale=True, skip_flds=['SK_ID_CURR'])
-
-    # trim extra NA flags
-    mapper.features = [f for f in mapper.features if f[0][0][-3:] != '_na']
-    mapper.built_features = [f for f in mapper.built_features if f[0][0][-3:] != '_na']
-    mapper.transformed_names_ = [f for f in mapper.transformed_names_ if f[-3:] != '_na']
-    df_numeric = df_numeric.iloc[:, :len(df.columns) - 1]
-    df_numeric['TARGET'] = y
-
-    # df_numeric: processed dataframe
-    # cat_size, cat_map: for CAT type data
-    # nas, mapper: for CONT type data
-    return df_numeric, cat_size, cat_map, nas, mapper
-
-
-def load_structured_test_data(df, cat_map, nas, mapper):
-    df = df[df['CODE_GENDER'] != 'XNA']
-
-    y = None
-    if 'TARGET' in df:
-        y = np.array(df['TARGET'])
-        df = df.drop('TARGET', axis=1)
-
-    for name, index_map in cat_map.items():
-        df[name] = df[name].map(index_map)
-
-    df_numeric, _, _, _ = proc_df(df, do_scale=True, skip_flds=['SK_ID_CURR'],\
-                                na_dict=nas, mapper=mapper)
-
-    # trim extra NA flags
-    df_numeric = df_numeric.iloc[:, :len(df.columns) - 1]
-    if y is not None: df_numeric['TARGET'] = y
-
-    return df_numeric
-
 
 class DNNTorch(BaseModel):
+    '''
+    Implements a Multi-input neural network for tabular data classification task
+    '''
     @staticmethod
     def get_knob_config():
         return {
@@ -257,69 +207,67 @@ class DNNTorch(BaseModel):
         }
 
     def __init__(self, **knobs):
-        super().__init__(**knobs)
-        self._knobs = knobs
+        self.__dict__.update(knobs)
 
-    def train(self, dataset_url, **kwargs):
-        # modify print function to rafiki mode
-        epoch = self._knobs.get('epoch')
-        lr = self._knobs.get('learning_rate')
+    def train(self, dataset_url, features=None, target=None, exclude=None, **kwargs):
 
-        utils.logger.define_plot('Loss Over Epochs',
-                                 ['train_loss', 'validate_loss', 'roc_auc_score'],
-                                 x_axis='epoch')
+        utils.logger.define_plot(
+            'Loss Over Epochs',
+            ['train_loss', 'validate_loss', 'roc_auc_score'],
+            x_axis='epoch')
 
+        # Record features & target
+        self._features = features
+        self._target = target
+
+        # Load CSV file as pandas dataframe
+        df = pd.read_csv(dataset_url, index_col=0)
+        df = df.drop(exclude, axis=1)
+
+        # Optional: Remove 4 applications with XNA CODE_GENDER (train set)
+        df = df[df['CODE_GENDER'] != 'XNA']
+
+        # Extract X & y from dataframe
+        (X, y) = self._extract_xy(df)
+        # Encode categorical features
+        X = self._encoding_categorical_type(X)
+        # other preprocessing
+        df_train = self._preprocessing(X)
+
+        # build model, set optimizer and loss function
+        self._build_model(self.layer_dim)
         device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+        self._model = self._model.to(device)
+        criterion = F.nll_loss
+        optimizer = torch.optim.SGD(self._model.parameters(),
+                                    lr=self.learning_rate, momentum=0.9)
 
-        # prepare dataset
-        # the file of dataset_url should be in CSV format and compatible for pandas
-        df_train, cat_size, cat_map, nas, mapper = \
-            load_structured_train_data(pd.read_csv(dataset_url, index_col=0))
-        self.raw_columns = ['SK_ID_CURR'] + list(df_train.columns)
-        self.cat_size = cat_size
-        self.cat_map = cat_map
-        self.nas = nas
-        self.mapper = mapper
-
-
-        y = np.array(df_train['TARGET'])
-        df_train.drop('TARGET', axis=1, inplace=True)
-        self.n_cont = len(df_train.columns) - len(cat_size)
-
-        df_to_nn_train, df_to_nn_valid, y_train, y_valid = train_test_split(
-            df_train, y, test_size=0.2, random_state=23, stratify=y)
-
-        tarin_loader = DataLoader(HomeCreditData(df_to_nn_train, y_train, cat_size),
+        # prepare train/valid dataset
+        cat_list = [n for n in self._encoding_dict]
+        x_train, x_valid, y_train, y_valid = train_test_split(df_train, y,
+                                    test_size=0.2, random_state=23, stratify=y)
+        tarin_loader = DataLoader(HomeCreditData(x_train, y_train, cat_list),
                                   batch_size=1000, num_workers=0, shuffle=True)
-        valid_loader = DataLoader(HomeCreditData(df_to_nn_valid, y_valid, cat_size),
+        valid_loader = DataLoader(HomeCreditData(x_valid, y_valid, cat_list),
                                   batch_size=5000, num_workers=0, shuffle=True)
-
-        del df_to_nn_train, df_to_nn_valid, y_train, y_valid, df_train, y
+        del x_train, x_valid, y_train, y_valid, df_train, y
         gc.collect()
 
-        # build model
-        self._build_model(cat_size)
-        self._model = self._model.to(device)
+        # start training
+        for i in range(self.epoch):
 
-        optimizer = torch.optim.SGD(self._model.parameters(), lr=lr, momentum=0.9)
-        self.criterion = F.nll_loss
-
-        train_losses = []
-        valid_losses = []
-        valid_scores = []
-
-        for i in range(epoch):
+            train_losses = []
+            valid_losses = []
+            valid_scores = []
 
             # training progress
             self._model.train()
             for idx, (x_cat, x_cont, y) in enumerate(tarin_loader):
-                x_cat = x_cat.to(device)
-                x_cont = x_cont.to(device)
-                y = y.to(device)
+                x_cat, x_cont, y = map(lambda x: x.to(device), [x_cat, x_cont, y])
 
                 optimizer.zero_grad()
                 output = self._model(x_cat, x_cont)
-                loss = self.criterion(output, y)
+                loss = criterion(output, y)
                 train_losses.append(loss)
 
                 loss.backward()
@@ -333,13 +281,11 @@ class DNNTorch(BaseModel):
             # validating progress
             self._model.eval()
             for idx, (x_cat, x_cont, y) in enumerate(valid_loader):
-                x_cat = x_cat.to(device)
-                x_cont = x_cont.to(device)
-                y = y.to(device)
+                x_cat, x_cont, y = map(lambda x: x.to(device), [x_cat, x_cont, y])
 
                 with torch.no_grad():
                     output = self._model(x_cat, x_cont)
-                    loss = self.criterion(output, y)
+                    loss = criterion(output, y)
 
                 score = roc_auc_score(y.cpu(), output[:, 1].cpu())
                 valid_losses.append(loss)
@@ -355,31 +301,34 @@ class DNNTorch(BaseModel):
                              roc_auc_score=valid_scores[-1],
                              epoch=i)
 
-            train_losses.clear()
-            valid_losses.clear()
-            valid_scores.clear()
-
     def evaluate(self, dataset_url):
-        device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-        df_eval = load_structured_test_data(
-                                pd.read_csv(dataset_url, index_col=0), \
-                                self.cat_map, self.nas, self.mapper)
+        # Load CSV file as pandas dataframe
+        df = pd.read_csv(dataset_url, index_col=0)
 
-        y = np.array(df_eval['TARGET'])
-        df_eval.drop('TARGET', axis=1, inplace=True)
+        # Optional: Remove 4 applications with XNA CODE_GENDER (train set)
+        df = df[df['CODE_GENDER'] != 'XNA']
 
-        eval_loader = DataLoader(HomeCreditData(df_eval, y, self.cat_size),
-                                 batch_size=10000, shuffle=False)
+        # Extract X & y from dataframe
+        (X, y) = self._extract_xy(df)
+        # Encode categorical features
+        X = self._features_mapping(X)
+        # other preprocessing
+        df_eval = self._preprocessing(X)
 
+        # prepare valid dataset
+        cat_list = [n for n in self._encoding_dict]
+        eval_loader = DataLoader(HomeCreditData(df_eval, y, cat_list),
+                                 batch_size=1000, num_workers=0, shuffle=False)
         del df_eval, y
         gc.collect()
 
+        device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
         eval_scores = []
         self._model.eval()
+
         for idx, (x_cat, x_cont, y) in enumerate(eval_loader):
-            x_cat = x_cat.to(device)
-            x_cont = x_cont.to(device)
-            y = y.to(device)
+            x_cat, x_cont, y = map(lambda x: x.to(device), [x_cat, x_cont, y])
+
             with torch.no_grad():
                 output = self._model(x_cat, x_cont)
 
@@ -391,19 +340,28 @@ class DNNTorch(BaseModel):
     def predict(self, queries):
 
         df = pd.DataFrame(queries)
-        df.columns = self.raw_columns[:-1]  # remove TARGET
-        df_test = load_structured_test_data(df, self.cat_map, self.nas, self.mapper)
 
-        test_loader = DataLoader(HomeCreditData(df_test, None, self.cat_size),
-                                    batch_size=5000, shuffle=False)
+        # Extract X & y from dataframe
+        (X, y) = self._extract_xy(df)
+        # Encode categorical features
+        X = self._features_mapping(X)
+        # other preprocessing
+        df_test = self._preprocessing(X)
+
+        # prepare valid dataset
+        cat_list = [n for n in self._encoding_dict]
+        test_loader = DataLoader(HomeCreditData(df_test, y, cat_list),
+                                 batch_size=1000, num_workers=0, shuffle=False)
+        del df_test, y
+        gc.collect()
 
         device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
         prediction = []
         self._model.eval()
+
         for idx, (x_cat, x_cont) in enumerate(test_loader):
-            # assert y is None
-            x_cat = x_cat.to(device)
-            x_cont = x_cont.to(device)
+            x_cat, x_cont = map(lambda x: x.to(device), [x_cat, x_cont])
+
             with torch.no_grad():
                 output = self._model(x_cat, x_cont)
 
@@ -416,7 +374,7 @@ class DNNTorch(BaseModel):
 
     def dump_parameters(self):
         params = {}
-        # Save model parameters
+        # Save network parameters
         with tempfile.NamedTemporaryFile() as tmp:
             # Save whole model to temp h5 file
             self._model.cpu()
@@ -425,26 +383,24 @@ class DNNTorch(BaseModel):
             with open(tmp.name + '.model', 'rb') as f:
                 h5_model_bytes = f.read()
 
-            model_config = pickle.dumps([self.cat_size, self.n_cont, self._knobs.get('layer_dim')])
-            data_config = pickle.dumps([self.raw_columns, self.cat_map, self.nas, self.mapper])
+        data_config_bytes = pickle.dumps([
+            self.layer_dim, self._encoding_dict, self._features, self._target,
+            self._nas, self._mapper
+        ])
 
-            params['h5_model_base64'] = base64.b64encode(h5_model_bytes).decode('utf-8')
-            params['model_config'] = base64.b64encode(model_config).decode('utf-8')
-            params['data_config'] = base64.b64encode(data_config).decode('utf-8')
+        params['h5_model_base64'] = base64.b64encode(h5_model_bytes).decode('utf-8')
+        params['data_config_base64'] = base64.b64encode(data_config_bytes).decode('utf-8')
 
         return params
 
     def load_parameters(self, params):
         # Load model parameters
         h5_model_base64 = params.get('h5_model_base64', None)
-        model_config_byte = params.get('model_config', None)
-        data_config_byte = params.get('data_config', None)
+        data_config_base64 = params.get('data_config_base64', None)
 
-        model_config = pickle.loads(base64.b64decode(model_config_byte.encode('utf-8')))
-        self.cat_size, self.n_cont, self._knobs['layer_dim'] = model_config
-
-        data_config = pickle.loads(base64.b64decode(data_config_byte.encode('utf-8')))
-        self.raw_columns, self.cat_map, self.nas, self.mapper = data_config
+        data_config_bytes = base64.b64decode(data_config_base64.encode('utf-8'))
+        self.layer_dim, self._encoding_dict, self._features, self._target, \
+            self._nas, self._mapper = pickle.loads(data_config_bytes)
 
         with tempfile.NamedTemporaryFile() as tmp:
             # Convert back to bytes & write to temp file
@@ -452,19 +408,69 @@ class DNNTorch(BaseModel):
             with open(tmp.name + '.model', 'wb') as f:
                 f.write(h5_model_bytes)
 
-            self._build_model(self.cat_size)
+            self._build_model(self.layer_dim)
             # Load model from temp file
             self._model.load_state_dict(torch.load(tmp.name + '.model'))
 
         device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
         self._model = self._model.to(device)
 
-    def _build_model(self, cat_size):
-        emb_sizes = [(c, min(50, (c + 1) // 2)) for _, c in cat_size]
-        layer_dim = self._knobs.get('layer_dim')
+    def _extract_xy(self, data):
+        if self._target is None:
+            self._target = 'TARGET'
+        y = data[self._target] if self._target in data else None
+
+        if self._features is None:
+            X = data.drop(self._target, axis=1)
+            self._features = list(X.columns)
+        else:
+            X = data[self._features]
+
+        return (X, y)
+
+    def _encoding_categorical_type(self, df):
+        # Apply label encoding for those categorical columns
+        cat_cols = [col for col in df.columns if df[col].dtype == 'object']
+        # deal with NaN and make index encoding
+        encoding_dict = {}
+        for name in cat_cols:
+            df[name] = df[name].astype('category').cat.as_ordered()
+            cat_codes = dict(enumerate(df[name].cat.categories))
+            # make way for NaN (index:0)
+            encoding_dict[name] = {cat: code + 1 for code, cat in cat_codes.items()}
+        self._encoding_dict = encoding_dict
+        return df
+
+    def _preprocessing(self,df):
+        # encoding, filling NaN and apply scaling on data
+        nas = self._nas if hasattr(self, '_nas') else None
+        mapper = self._mapper if hasattr(self, '_mapper') else None
+        df_numeric, _, nas, mapper = proc_df(df, do_scale=True, na_dict=nas, mapper=mapper)
+
+        # trim extra NA flag columns
+        mapper.features = [f for f in mapper.features if f[0][0][-3:] != '_na']
+        mapper.built_features = [f for f in mapper.built_features if f[0][0][-3:] != '_na']
+        mapper.transformed_names_ = [f for f in mapper.transformed_names_ if f[-3:] != '_na']
+        df_numeric = df_numeric.iloc[:, :len(df.columns)]
+
+        self._nas, self._mapper = nas, mapper
+        return df_numeric
+
+    def _features_mapping(self, df):
+        # Encode the categorical features with pre saved encoding dict
+        cat_cols = [col for col in df.columns if df[col].dtype == 'object']
+        df_tmp = df.copy()
+        for name in cat_cols:
+            df_tmp[name] = df[name].map(self._encoding_dict[name]).fillna(0)
+        df = df_tmp
+        return df
+
+    def _build_model(self, layer_dim):
+        emb_sizes = [(len(types) + 1, round(len(types)**0.5))
+                     for _, types in self._encoding_dict.items()]
         self._model = MixedInputModel(
             emb_sizes=emb_sizes,
-            n_cont=self.n_cont,
+            n_cont=len(self._features) - len(self._encoding_dict),
             emb_drop=0.05,
             out_size=2,
             sizes=[layer_dim * 2, layer_dim, layer_dim],
@@ -476,19 +482,21 @@ class DNNTorch(BaseModel):
 
 
 if __name__ == '__main__':
-    curpath = os.path.join(os.environ['HOME'], 'Desktop', 'rafiki-master')
+    curpath = os.path.join(os.environ['HOME'], 'rafiki')
     os.environ.setdefault('WORKDIR_PATH', curpath)
     os.environ.setdefault('PARAMS_DIR_PATH', os.path.join(curpath, 'params'))
 
-    train_set_url = os.path.join(curpath, 'input', 'application_train_index.csv')
+    train_set_url = os.path.join(curpath, 'data', 'application_train_index.csv')
     valid_set_url = train_set_url
-    test_set_url = os.path.join(curpath, 'input', 'application_test_index.csv')
-    test_queries = pd.read_csv(test_set_url, index_col=0).values.tolist()
+    test_set_url = os.path.join(curpath, 'data', 'application_test_index.csv')
+
+    test_queries = pd.read_csv(test_set_url, index_col=0).iloc[:5]
+    test_queries = json.loads(test_queries.to_json(orient='records'))
 
     test_model_class(
         model_file_path=__file__,
         model_class='DNNTorch',
-        task=None,
+        task='TABULAR_CLASSIFICATION',
         dependencies={
             ModelDependency.TORCH: '1.3.1',
             ModelDependency.SCIKIT_LEARN: '0.21.3',
@@ -497,5 +505,9 @@ if __name__ == '__main__':
         },
         train_dataset_path=train_set_url,
         val_dataset_path=valid_set_url,
+        train_args={
+            'target': 'TARGET',
+            'exclude': ['SK_ID_CURR'],
+        },
         queries=test_queries,
     )
