+++ examples/models/tabular_classification/LightGBM.py
@@ -44,35 +44,6 @@ warnings.simplefilter(action='ignore', category=FutureWarning)
     'application_{train|test}.csv' of this competition as dataset.
 '''
 
-###################################
-# preprocessing for structured data
-###################################
-
-def preprocess_data(df):
-    # Optional: Remove 4 applications with XNA CODE_GENDER (train set)
-    df = df[df['CODE_GENDER'] != 'XNA']
-
-    # Categorical features with Binary encode (0 or 1; two categories)
-    for bin_feature in ['CODE_GENDER', 'FLAG_OWN_CAR', 'FLAG_OWN_REALTY']:
-        df[bin_feature], uniques = pd.factorize(df[bin_feature])
-
-    for col in df.columns:
-        if df[col].dtype == 'object':
-            df[col] = df[col].astype('category')
-
-    # NaN values for DAYS_EMPLOYED: 365.243 -> nan
-    df['DAYS_EMPLOYED'].replace(365243, np.nan, inplace= True)
-    # Some simple new features (percentages)
-    df['DAYS_EMPLOYED_PERC'] = df['DAYS_EMPLOYED'] / df['DAYS_BIRTH']
-    df['INCOME_CREDIT_PERC'] = df['AMT_INCOME_TOTAL'] / df['AMT_CREDIT']
-    df['INCOME_PER_PERSON'] = df['AMT_INCOME_TOTAL'] / df['CNT_FAM_MEMBERS']
-    df['ANNUITY_INCOME_PERC'] = df['AMT_ANNUITY'] / df['AMT_INCOME_TOTAL']
-    df['PAYMENT_RATE'] = df['AMT_ANNUITY'] / df['AMT_CREDIT']
-    # del test_df
-    gc.collect()
-    return df
-
-
 class LightGBM(BaseModel):
 
     @staticmethod
@@ -86,45 +57,46 @@ class LightGBM(BaseModel):
         }
 
     def __init__(self, **knobs):
-        super().__init__(**knobs)
-        self._knobs = knobs
+        self.__dict__.update(knobs)
 
-    def train(self, dataset_url, **kwargs):
-
-        self.data_addr = dataset_url
+    def train(self, dataset_url, features=None, target=None, exclude=None, **kwargs):
         utils.logger.define_plot('Loss Over Epochs',
                                  ['loss', 'early_stop_val_loss'],
                                  x_axis='epoch')
-        target='TARGET'
-        df = pd.read_csv(dataset_url, sep=',', index_col=0)
-        train_df = preprocess_data(df)
 
-        self.raw_columns = list(df.columns)
-        del df
+        self._features = features
+        self._target = target
+
+        df = pd.read_csv(dataset_url, index_col=0)
+        df = df.drop(exclude, axis=1)
+
+        # Optional: Remove 4 applications with XNA CODE_GENDER (train set)
+        df = df[df['CODE_GENDER'] != 'XNA']
+
+        # Extract X & y from dataframe
+        (X, y) = self._extract_xy(df)
+        # Encode categorical features
+        X = self._encoding_categorical_type(X)
+        # other preprocessing
+        df_train = self._preprocessing(X)
 
         # Cross validation model
         folds = KFold(n_splits= 10, shuffle=True)
-        # Create arrays and dataframes to store results
-        oof_preds = np.zeros(train_df.shape[0])
-        feats = [f for f in train_df.columns if f not in [target]]
         flag=0
-        for n_fold, (train_idx, valid_idx) in enumerate(folds.split(train_df[feats], train_df[target])):
-
-            train_x, train_y = train_df[feats].iloc[train_idx], train_df[target].iloc[train_idx]
-            valid_x, valid_y = train_df[feats].iloc[valid_idx], train_df[target].iloc[valid_idx]
-            lgb_train = lgb.Dataset(train_x, train_y,)
-            lgb_valid = lgb.Dataset(valid_x, valid_y,)
+        for n_fold, (train_idx, valid_idx) in enumerate(folds.split(X, y)):
+            lgb_train = lgb.Dataset(X.iloc[train_idx], y.iloc[train_idx],)
+            lgb_valid = lgb.Dataset(X.iloc[valid_idx], y.iloc[valid_idx],)
             params = {
                 'boosting_type': 'gbdt',
                 'objective': 'binary',
                 'metric': 'cross_entropy',
                 'nthread': 4,
                 'n_estimators': 10,
-                'learning_rate': 0.02,
-                'num_leaves': self._knobs['num_leaves'],
-                'colsample_bytree': self._knobs['colsample_bytree'],
-                'subsample': self._knobs['subsample'],
-                'max_depth': self._knobs['max_depth'],
+                'learning_rate': self.learning_rate,
+                'num_leaves': self.num_leaves,
+                'colsample_bytree': self.colsample_bytree,
+                'subsample': self.subsample,
+                'max_depth': self.max_depth,
                 'verbose': -1,
             }
 
@@ -141,25 +113,33 @@ class LightGBM(BaseModel):
             flag+=1
 
     def evaluate(self, dataset_url):
-        target='TARGET'
-        df = pd.read_csv(self.data_addr,sep=',', index_col=0).iloc[:1000]
-        df = preprocess_data(df)
+        df = pd.read_csv(dataset_url, index_col=0)
+        
+        # Optional: Remove 4 applications with XNA CODE_GENDER (train set)
+        df = df[df['CODE_GENDER'] != 'XNA']
 
-        oof_preds = np.zeros(df.shape[0])
-        feats=[f for f in df.columns if f not in [target]]
-        valid_x=df[feats]
-        valid_y=df[target]
-        oof_preds=self._model.predict(valid_x)
-        return roc_auc_score(valid_y,oof_preds)
+        # Extract X & y from dataframe
+        (X, y) = self._extract_xy(df)
+        # Encode categorical features, no need mapping features for this model
+        X = self._encoding_categorical_type(X)
+        # other preprocessing
+        df_train = self._preprocessing(X)
 
+        # oof_preds = np.zeros(df.shape[0])
+        oof_preds=self._model.predict(X)
+        return roc_auc_score(y,oof_preds)
 
     def predict(self, queries):
         df = pd.DataFrame(queries)
-        self.raw_columns.remove('TARGET')
-        df.columns = self.raw_columns
-        test_df = preprocess_data(df)
 
-        predictions=self._model.predict(test_df)
+        # Extract X & y from dataframe
+        (X, y) = self._extract_xy(df)
+        # Encode categorical features
+        X = self._encoding_categorical_type(X)
+        # other preprocessing
+        df_train = self._preprocessing(X)
+
+        predictions=self._model.predict(X)
         predicts = []
         for prediction in predictions:
             predicts.append(prediction)
@@ -170,51 +150,90 @@ class LightGBM(BaseModel):
 
     def dump_parameters(self):
         params = {}
-
         # Save model parameters
         with tempfile.NamedTemporaryFile() as tmp:
             # Save whole model to temp h5 file
-
             self._model.save_model(tmp.name)
-
             # Read from temp h5 file & encode it to base64 string
             with open(tmp.name, 'rb') as f:
                 h5_model_bytes = f.read()
 
-            data_config = pickle.dumps(self.raw_columns)
-            params['h5_model_base64'] = base64.b64encode(h5_model_bytes).decode('utf-8')
-            params['data_config'] = base64.b64encode(data_config).decode('utf-8')
+        data_config_bytes = pickle.dumps([self._features, self._target])
+
+        params['h5_model_base64'] = base64.b64encode(h5_model_bytes).decode('utf-8')
+        params['data_config_base64'] = base64.b64encode(data_config_bytes).decode('utf-8')
 
         return params
 
     def load_parameters(self, params):
         # Load model parameters
         h5_model_base64 = params.get('h5_model_base64', None)
-        data_config_byte = params.get('data_config', None)
-        self.raw_columns = pickle.loads(base64.b64decode(data_config_byte.encode('utf-8')))
+        data_config_base64 = params.get('data_config_base64', None)
+
+        data_config_bytes = base64.b64decode(data_config_base64.encode('utf-8'))
+        self._features, self._target = pickle.loads(data_config_bytes)
 
         with tempfile.NamedTemporaryFile() as tmp:
             # Convert back to bytes & write to temp file
             h5_model_bytes = base64.b64decode(h5_model_base64.encode('utf-8'))
             with open(tmp.name, 'wb') as f:
                 f.write(h5_model_bytes)
-
             # Load model from temp file
             self._model =lgb.Booster(model_file=tmp.name)
+
+    def _extract_xy(self, data):
+        if self._target is None:
+            self._target = 'TARGET'
+        y = data[self._target] if self._target in data else None
+
+        if self._features is None:
+            X = data.drop(self._target, axis=1)
+            self._features = list(X.columns)
+        else:
+            X = data[self._features]
+
+        return (X, y)
+
+    def _encoding_categorical_type(self, df):
+        # Categorical features with Binary encode (0 or 1; two categories)
+        for bin_feature in ['CODE_GENDER', 'FLAG_OWN_CAR', 'FLAG_OWN_REALTY']:
+            df[bin_feature], uniques = pd.factorize(df[bin_feature])
+
+        for col in df.columns:
+            if df[col].dtype == 'object':
+                df[col] = df[col].astype('category')
+        return df
+
+    def _preprocessing(self,df):
+        # NaN values for DAYS_EMPLOYED: 365.243 -> nan
+        df['DAYS_EMPLOYED'].replace(365243, np.nan, inplace= True)
+        # Some simple new features (percentages)
+        df['DAYS_EMPLOYED_PERC'] = df['DAYS_EMPLOYED'] / df['DAYS_BIRTH']
+        df['INCOME_CREDIT_PERC'] = df['AMT_INCOME_TOTAL'] / df['AMT_CREDIT']
+        df['INCOME_PER_PERSON'] = df['AMT_INCOME_TOTAL'] / df['CNT_FAM_MEMBERS']
+        df['ANNUITY_INCOME_PERC'] = df['AMT_ANNUITY'] / df['AMT_INCOME_TOTAL']
+        df['PAYMENT_RATE'] = df['AMT_ANNUITY'] / df['AMT_CREDIT']
+        return df
+
+    def _features_mapping(self, df):
+        pass
 
     def _build_model(self):
         pass
 
 
 if __name__ == '__main__':
-    curpath = os.path.join(os.environ['HOME'], 'Documents', 'Github', 'rafiki')
+    curpath = os.path.join(os.environ['HOME'], 'rafiki')
     os.environ.setdefault('WORKDIR_PATH', curpath)
     os.environ.setdefault('PARAMS_DIR_PATH', os.path.join(curpath, 'params'))
 
     train_set_url = os.path.join(curpath, 'data', 'application_train_index.csv')
     valid_set_url = train_set_url
     test_set_url = os.path.join(curpath, 'data', 'application_test_index.csv')
-    df_test = pd.read_csv(test_set_url, index_col=0)
+
+    test_queries = pd.read_csv(test_set_url, index_col=0).iloc[:5]
+    test_queries = json.loads(test_queries.to_json(orient='records'))
+
 
     test_model_class(
         model_file_path=__file__,
@@ -226,6 +245,9 @@ if __name__ == '__main__':
         },
         train_dataset_path=train_set_url,
         val_dataset_path=valid_set_url,
-        # budget={'MODEL_TRAIL_COUNT': 1},
-        queries=df_test.values.tolist(),
+        train_args={
+            'target': 'TARGET',
+            'exclude': ['SK_ID_CURR'],
+        },
+        queries=test_queries,
     )
